requirements.txt
openneuro-py
mne
mne-bids
tensorflow
numpy
pandas
matplotlib
seaborn
scikit-learn